Compiler Package Changes to support open source build.
How it works, the Workspace uses the ExecutingAssembly path to determine in which hive the package is running. It passes this as a Build Propert to msbuild.

Whenever the compiler package is loaded by vs it scans the bs/vwd/wd and vswinexpress directories to find all of the hives and writes out a when clause for each found hive.  The when clauses verify the existence of the roslyntoolpath directory before setting it.  The otherwise clause is the 12.0 hive for the app which has the directory.

 (changeset 1209363)