Bug fix for 907771: codegen issue around fixed blocks in unoptimized code.
* the code that was injecting indirect returns had unreliable logic when detecting the appropriate block.
We could end up injecting return in the middle of the method body and the rest would get dropped as unreachable.

* added debug visualization support for BasicBlock and ILBuilder.
We used to have visualization here shared with IL dumping, but it gradually migrated out to the test utilities.
It has been a while since we had to fix a bug in codegen, but now, I realize that it is fairly hard to debug this kind of issues without visualizers.
 (changeset 1209393)