DevDiv #906494: Overload resolution diagnostic reporting
We were throwing unreachable when we fell out the bottom of some complicated error handling code and Tomas hit it while dogfooding.  We were unable to determine the path that resulted in the crash, but there were enough problems that a more comprehensive cleanup seemed desirable.  The catch-all diagnostic is now checked last and actually serves as a catch-all.

Miscellaneous improvements:
1) Eliminated MemberCouldTakeArgumentCount which was trying to approximate overload resolution.
2) Standardized mechanism for iterating of list of candidates.
3) Eliminated MemberResolutionKind.BadGenericArity.  Comments explicitly stated that it was never used.

CR: AlekseyT (changeset 1211059)