Better support for reading ProjectReference items in MSBuildWorkspace.
When MSBuildWorkspace is reading .csproj/.vbproj files, it now doesn't crash if the <Project> metadata item (GUID) on a <ProjectReference> item is not set (there are some project references out there that don't specify a GUID). Since VS allows such project references, we should too.

Also, if <ReferenceOutputAssembly> is set to false, we will now just skip the reference. Earlier we were ignoring this metadata and always adding a reference. (changeset 1211532)