We have types Union<T0,T1> and Union<T0,T1,T2> in the compiler layer. They're used in just one place each - in the various "DescendantNodes" routines. The allocations for these unions are wasteful because the size of the combined structure is the sum of the sizes each part plus one for a "discriminator" byte.
The allocations were showing up as ~1.5% of the C# goldilocks typing scenario.
This change replaces the "Stack of Union" with a "UnionStack". The Union stack is actually multiple stacks in one - one for each type and a disciminator "master" stack.
We can also make use of object pooling to further reduce allocations. (changeset 1213710)