Fix a bug where MSBuildWorkspace wasn't signing projects and thus InternalsVisibleTo was broken when opening solutions outside of Visual Studio.
In an ongoing effort to make the compiler portable we moved out a StrongNameProvider and made it a part of CompilationOptions. When MSBuildWorkspace was creating CompilationOptions, it didn't pass a StrongNameProvider and so when loading signed projects all we were getting is a diagnostic:

error CS7027: Error signing output with public key from file 'C:\...' -- Assembly signing not supported.

We now pass better defaults when creating CompilationOptions for loading MSBuild projects via MSBuildWorkspace.

This checkin also adds a unit-test to test IVT to signed projects in MSBuildWorkspace. To make it happen I introduce SolutionGeneration.cs, an API that can be used by unit-tests to construct MSBuild projects and solutions. This eliminates the need to bake in canned projects and solutions as resources in the test assembly.

If this unit-testing approach proves comfortable and maintainable, I will be gradually removing canned resources from the test assembly and replacing them with declarative solution construction API calls. Here's an example:

            var solution = Solution(
                Project(
                    ProjectName("Project1"),
                    Sign,
                    Document(string.Format(
@"using System.Runtime.CompilerServices;
[assembly:InternalsVisibleTo(""Project2, PublicKey={0}"")]
class C1
{{
}}", PublicKey))),
                Project(
                    ProjectName("Project2"),
                    Sign,
                    ProjectReference("Project1"),
                    Document(@"class C2 : C1 { }")));

This is a fully self-contained description of a solution sufficient to reproduce the bug fixed by this checkin. When writing a unit-test, no need to fabricate a sample solution and add it as resources. When reading a unit-test, no need to go somewhere look something up - it's all right here, all relevant details present, all irrelevant details (such as document names) omitted and filled out with reasonable defaults. This API will be further fleshed out as needed, as requirements arise. (changeset 1214118)