This is the toolset changes to enable Open source developers to debug the results of their compiler hacking efforts.
How it works:
Unfortunately we can't do VSIX magic or binding redirect magic.  However, that leaves a pretty easy to understand solution.
The Opensource developers will work in the defaultmain hive, their compiled product will be installed in the roslyndev hive.
During the build we will overwrite the binaries in their deployed locations with the result of the build.
They select CompilerPackage as their start package and press F5 this will build their product and deploy it to the start a new instance of Visual Studio using the Roslyndev hive and their new code.

How to use it:
There is a one time initialization prepare.bat, this must be run as an administrator, it creates the RoslynDev hive, Initializes and does sn - Vr on each Roslyn Assembly that the open source project builds

to undo this, run prepare.bat with the u parameter.

Once you have run prepare.bat set a global environment variable "OpenSourceDebug" to true just load up the roslyn solution and press F5, the build will put the right bits at the right place in the hive.  If the variable is not set then the normal build occurs and the RoslynHive will not be modified.

There really isn't that much to it.

Please Note I have hard coided the version # to 0.6.0.0 when building For OpenSourceDebug, otherwise the

Kevin

 (changeset 1216751)