This is a minimal fix for tests VisualBasic.UnitTests.CommandLineTests.BinaryFile and CSharp.UnitTests.CommandLineTests.BinaryFileErrorTest and the issue that they surface.
We have a heuristic in EncodedStringText.DecodeIfNotBinary() that checks if a non-UTF8-encoded source file passed as argument for command line compiler is a binary file, and quickly fails decoding in this case. However, this heuristic only ran in case system default encoding was Windows-1252.

The perfect fix would be to extend DecodeIfNotBinary, to apply heuristic for any possible encoding, but this requires substantial changes to support multi-byte encodings (Windows-1252 is a single-byte one, so a read buffer can be decoded on each Stream.Read).

But I'll leave the correct fix till after Build, meanwhile making the most isolated change that would fix the tests and apply heuristic to binary files under any system locale, which would be to default to Windows-1252 for any non-UTF8 source file, if the user has not specified code page explicitly. (changeset 1219639)