DevDiv #913556: Crash while emitting large string switch without string.Chars
Instead of trying to have emit repeat the checks performed during lowering, just have it consume the results of those checks (i.e. whether or not the string switch helper method was actually synthesized).

VB already had a test documenting that this scenario is unsupported (appropriate errors are reported). (changeset 1220615)