Bugfix 881746
The change is, when a syntax is simplified into its Alias at Attribute Context (i.e. within Attribute) we make sure that we do not remove the suffix "attribute" from the  new Identifier name because the new identifier is now referencing the Alias.

Having the Identifiers meaning the same with "FooAttribute" and "Foo", where "FooAttribute" inherits System.Attribute, in the Attribute context does not apply for Aliases that has "Attribute" suffix. (changeset 1222441)