Reverting the changes that introduced asynchronous reading of source files in batch compile scenarios
It appeared that the change did not result in noticeable improvements in IO throughput in scenarios where it matters.
That is most likely because async reading would be beneficial in a case of reading a small number of large files, while we typically have the opposite - large number of small files.

On the other hand, on not IO-bound scenarios, the change resulted in parsing throughput regressions since it can cause oversubscription of threadpool.
 (changeset 1226132)