This fixes EncodedStringText to apply the binary file detection heuristic to text in any encoding, and also correctly pick Encoding.Default (system locale-dependent) instead of Windows-1252, when encoding is not specified.
Previously, the heuristic only applied if system default codepage was 1252 (default for ENU), thus when a random binary file was passed to command-line compiler as a source code argument on a system with non-ENU locale, a huge number of diagnostics were issued and compiler crashed with an OutOfMemoryException.

This was promptly patched before BUILD by always picking Windows-1252 if encoding was unspecified, thus introducing a breaking change to users on non-ENU systems - if they had non-Unicode sources with extended ASCII characters, their code could fail to compile when transitioning from native compiler to Roslyn.

This fix rolls back this change, and also changes the heuristic to rely on comparing decoded characters to NUL, instead of looking at encoded bytes. (The heuristic is to look for two consecutive NUL characters in decoded character stream, and if those are found, fail the decoding process.) The latter seemed to be applicable to vast majority of possible encodings, with a notable exception of UTF-16 and UTF-32, where two consecutive zero bytes could appear in a legitimate text file. Relying on characters allows us to not depend on the encoding choice.

As a slight downside we're now decoding the entire string (reader.ReadToEnd) before applying the heuristic, though this seems negligible compared to parsing/binding, which takes much more time - we're avoiding that for detected binary files.

On the other hand, we're now relying on trusted StreamReader to decode a string, instead of implementing buffered decoding logic ourselves, as we previously did in DecodeIfNotBinary, and which was prone to buffer boundary errors if we ever used it for multi-byte encodings (though we only called this method for Windows-1252).
 (changeset 1226686)