Added the first diagnostic analyzer (+tests) which will eventually analyze Roslyn.sln. It produces the following diagnostics: RS0001: Use SpecializedCollections.EmptyEnumerable<T>() instead of allocating an array of length 0 or using System.Linq.Enumerable.Empty<T>() RS0002: Use SpecializedCollections.SingletonEnumerable<T>() instead of allocating an array of length 1.
In order to get RS0002 completely correct, analyzers need to be able to do data flow analysis, so there are a few skipped tests for now.

In the process of implementing the unit tests, I also pulled the diagnostic analyzer test base into a separate project, DiagnosticsTestUtilities. I updated the analyzer test base to provide output formatted to be immediately usable in test cases, similar to the compiler diagnostics tests. (changeset 1229264)