Add a method to get all the files that participate in a ruleset.
The RuleSet type currently has a method, LoadEffectiveRuleSetFromFile, that processes a ruleset file and all of the other ruleset files it includes to create a final ruleset taking into account which files override which other files. However, there is no way to determine the complete set of files that participated in this process.

This change adds a GetEffectiveIncludesFromFile method to do just that. (changeset 1231203)