Bug Fix for Devdiv:817145.
Fixing bugs in type inferrer which were inadvertently introduced in Changeset 822325. The CS had changed the return type of a method from some T to IEnumerable<T> and some of the call sites weren't updated accordingly. These callers were using 'var' for type inference, so the build didn't break but the semantics had silently been changed.

Enabled an existing unit test for verifying that completion list builder and added an unit test for the typeinferrer. (changeset 1232655)