Emit with metadataOnly should not emit anonymous types
Compared to other synthetic types, anonymous types are special - they are visible through semantical models and therefore they are indirectly owned by the Compilation (not by the ModuleBuilder).
Since anonymous types are an artifact of method body binding, normally metadataOnly emit does not need to deal with anonymous types. However, if the same compilation is used for a complete (not metadata only) emit, the anonymous types materialize and become visible to the metadata-only emit too.
That is particularly inconvenient when different kinds of emit happen concurrently and metadata emit may try emitting incomplete set of anonymous types and cause strange crashes.

A simple solution to all these issues is to make metadata-only emit to pretend that anonymous types do not exist even if present in the compilation. (changeset 1233610)