Tweak SimplificationOptions to use separate feature names for per-language and global options.
Without this, the option service couldn't distinguish between the two, and so it was loading the feature-specific option serializers for non-feature-specific options. This not only was a waste of resources (the VB serializers don't need to be loaded if all you have is C# open), but also risked deadlocks due to the subtleties in how serializers in VS must be created. (changeset 1238958)