Fix a race condition in rule set validation.
When loading a .ruleset file we first validate it against a schema. However, we have a single XmlSchema object to do this, held in a static field. This means that if VBCSCompiler.exe is processing multiple build requests at once we may end up using this object simultaneously from multiple threads, and it isn't thread-safe.

The fix here is to create a new XmlSchema every time we need it. (changeset 1239493)