WarnAsError shouldn't turn info\hidden diagnostics into errors
When warnaserror is set for a project, we set the GeneralDiagnosticOption on CompilationOptions to be error and that turns any diagnostic that doesn't have a specific option set to an error. This means diagnostcs from simplify typename etc. get turned into errors and squiggles everything. The consensus from the diagnostic team meeting was this:
* When doing warn-as-error, don't raise severity for anything *below* warning (info or hidden).
* When doing suppress-all-warnings, don't lower severity for anything *above* warning (error).

Implementing this in C# and VB. Tests will follow.
 (changeset 1239863)