This fixes #90 (https://roslyn.codeplex.com/workitem/90).
There were issues with the PE headers generated by the compiler when emitting assemblies with no win32 resources (default behavior when you call Compilation.Emit).

If the size of a section is zero, then the relative virtual address should also be zero.  Previously, we would unconditionally set the address of the resource section to the end of the most recently preceding section that contained any data.  This was wrong in and of itself, but it also had the knock-on effect of breaking the logic in PeWriter.WriteRelocSection that is supposed to pad the last section to align with the file alignment boundary. (changeset 1240642)