Bugfix: Creating a C# Mercury project followed by a new VB Mercury project throws exception
If an open file is a linked file and that document or its project is unloaded, then we determine the new active context document for that linked file and switch the context. Following the same process during solution close may fail because the host document for the new active context document may have already been removed. There is no need to set the updated document context during solution close so we now skip this step in that case. (changeset 1251204)