Round-out the parts of the SemanticModel API that expose diagnostics.
Currently it is possible to call GetDiagnostics() to get all syntax, declaration, and method body diagnostics, or GetDeclarationDiagnostics() or GetMethodBodyDiagnostics() to get just declaration or method body diagnostics, respectively. However, there is no corresponding method to get just the syntax diagnostics. If you just want syntax warnings and errors you either have to bypass the SemanticModel or get all the diagnostics and try to separate syntax issues from semantic issues (and pay the cost of binding while you're at it).

This change fills the gap in the API by adding a GetSyntaxDiagnostics method.

 (changeset 1256304)