Ignore rule set <Include> elements that we don't know how to resolve.
The compilers cannot currently resolve the same range of <Include> elements as FxCop. Consider the following:

  <Include Path="MinimumRecommendedRules.ruleset" Action="Default" />

FxCop knows to go look for MinimumRecommendedRules.ruleset among the rule sets installed with VS, as well as in any user-defined locations. Roslyn lacks this ability, and so some projects that previously built will now fail with an error about the file not being found.

For the moment, just work around the issue by ignoring any included files that can't be found. (changeset 1261868)