The compiler used to use the metadata reference path to filter out duplicated references. Duplicate references were also filtered out based on assembly identity, if it the identity was strong. This change removes the former mechanism in favor of the latter modified to work for weak-named identities as well.
As a result the compiler will no longer report CS1704/BC32208 in cases where two metadata references passed to the compilation that have different paths but the same weak name and version. Based on SQM data these errors were rarely reported anyways. In addition it is hard to make the compilers report these errors when building from VS. So the value of these errors is minimal.

The goal of de-duplication is mainly to support #r scenarios, where application of #r on the "same" reference should be idempotent. The definition of "same" is not based on file paths any more but on simple name and version (or strong name if the assembly is signed).

Also removes FilePath from Location. The value was not well defined and thus problematic for metadata locations. (changeset 1270403)