removed assert since there is a chance where entry in the typemap might not be there yet in concurrent case
it is due to us using two concurrent maps, so it is possible for us to put one information in one map but not in the other one yet.

when we actually consume this binding object, we will never consume Record ang GetType/GetReader at the same time, so removing the assert should be safe. (changeset 1271843)