Remove dependency on XmlTextReader.
We used to use XmlTextReader to save allocations of underlying buffers while validating XML doc comments. XmlTextReader is a legacy non-portable class. However it has an advantage over a reader created via XmlReader.Create in the ability of resetting. The latter can't be reset and thus can't be pooled.

With this change, each time we need to validate a fragment of XML we "append" it to the underlying text stream and advance the reader pass the added fragment. By the end of the fragment validation we keep the reader open in a state that is ready for the next fragment validation unless the fragment was invalid, in which case we need to create a new XmlReader. (changeset 1278873)