Improve CompilationOptions equality checks
VS sometimes calls Workspace.OnCompilationOptionsChanged with a new CompilationOptions object that was created from the same inputs as the current CompilationOptions object.  However, since some classes were not overriding object.Equals, the solution was forked unnecessarily.  This change adds the missing implementations of Equals and GetHashCode.

Workspace.SetCurrentSolution is updated to not increment the WorkspaceVersion if the new solution is equivalent to the CurrentSolution.

Workspace.RaiseWorkspaceChangedEventAsync no longer fires events if the new solution is equivalent to the old solution. (changeset 1279344)