Bugfix 954386
Format Command Handler when deciding the range to format when selecting the first statement of a block/ first member of a type declaration does not include the '{' before the statement/member declaration.

This change includes the '{' in the above cases.

Eg 1:
class C
[|{ int s;|]
}

The range to be selected for formatting on typing ';' should be [|{ int s;|]. (changeset 1279830)