Fix for bug 967127: VBCSCompiler crashing while loading analyzers
Changeset 1272615 tweaked the analyzer loading code as follows: If Assembly.LoadFrom throws a ReflectionTypeLoadException, our current behavior is to log a compiler warning and skip the analyzer assembly. However, this exception might be caused by some unrelated missing types in referenced assemblies and it is possible to continue looking for analyzers from the loaded types in the assembly. This change tweaks the behavior to now log a separate warning for a ReflectionTypeLoadException, and continues working with only the loaded types.

However, the subsequent code that works with loaded types and tries to check if it implements IDiagnositcAnalyzer and has a DiagnosticAnalyzerAttribute might throw. Fix is to wrap this code within the existing try-catch for instantiating analyzers. (changeset 1280424)