Fixes error recovery for parsing of types when user has not typed an open brace. Consider following code
     namespace ns
     {
         class c

         //comment
     }

Prior to this fix, we would parse the close brace of the namespace as the close brace of the class (even when the open brace is missing). This quirk meant that features that need to generate code into the class (such as the implement interface feature) would do the wrong thing.

This fix changes the parser to synthesize missing open and close braces for the class and leaves the close brace that the user typed associated with the surrounding namespace. This in turn ensures that when the implement interface feature generates code, it can spit the missing braces for the class along with the interface methods. (changeset 1280826)