PEWriter metadata sizes fif.
The PEWriter used to calculate sizes of metadata tables, rows and columns 3 times during emit. Under certain circumstances the result of the first one was different from the other two which might have lead to writing invalid values to PE headers.
This change factors various sizes related to metadata tables to a separate class that is immutable, calculated once and then passed around.
 (changeset 1281677)