Passing non-code files and option to csc\vbc
Diagnostic analyzers need to be able to look at non-code files from a project (like web.config) to produce issues. They also need to be able to see project-level options. The IDiagnosticAnalyzer API will pass these files and options to the analyzers. So all analyzer hosts need to be able to pass these values through. This change adds the plumbing needed in csc\vbc by adding two switches /additionalfile and /option.

This is part 1 of the change to parse the commandline arguments. Passing to the analyzer will come in a follow up change. (changeset 1282218)