Implements parser support for the new #Enable / #Disable feature in VB along with changes required for lighting up following IDE features: classification, keyword recommender, formatting.
Speclet - http://roslyn.codeplex.com/discussions/543476. Feature is equivalent to the #pragma warning feature in C#.

ToDo: Changes to actually suppress / unsupress diagnostics in the compiler are NYI. IDE support for the new directives should be mostly complete except for keyword highlighting which I intend to look into in a subsequent change. (changeset 1283171)