Changes emit for synthetic methods backing lambda expressions to include dummy "this" parameter to match the number of parameters in delegate "Invoke" method.
When implementing static lambdas (those that do not have closures) C# has a choice of emitting an open or a closed delegate. In the latter case we can just enclose null and not use it in the method.
It appears that in a case of closed delegate, Invoke dispatch is considerably cheaper - no need to left-shift the arguments to adjust for missing "this".

== Example:

Consider code like the following

Func<int, int, int> adder = (x, y) => x + y;

Since the lambda does not need to enclose anything, we can implement it in two ways.

• Open delegate implementation (current)

        public static int Add(int x, int y)
        {
            return x + y;
        }

        Func<int, int, int> adderOpen = cls1.Add;

• Closed delegate implementation (also works and seems better)

        // does not really need to be an extension when in synthetic code
        public static int Add(this object dummy, int x, int y)
        {
            return x + y;
        }

        // enclose null
        Func<int, int, int> adderClosed = ((object)null).Add;

Crude experiments confirm that invoking adderClosed indeed costs noticeably less than adderOpen. (about 30% less in common scenarios like adding two ints)

This change adjusts codegen for static lambdas to use closed delegates.

 (changeset 1285058)