More robust command line compiler testing
Previously the command line tests were looking in various directories for csc.exe/vbc.exe. Instead, add a reference to csc/vbc project to the command line unit test assembly, so that we always test what we build and not depend on directories outside of Roslyn.
To make it easy to test the default .rsp file this change links it into the command line unit test project and includes it to resources. The tests then extract the .rsp from the resources and save it to a temp file for testing.
 (changeset 1285693)