As part of hooking up the MicroBuild signing task, we need to modify the Roslyn build targets to use the new task. I've made the logic conditional so that if Roslyn is being built in the MicroBuild environment, it will use the MicroBuild signing task, and if not, it will continue using the task that it is currently using today. Once all of Roslyn's official builds are using the MicroBuild process, this conditional logic can be removed along with the old task in Roslyn.Setup.BuildTasks.dll.
The MicroBuild template will pass in a SignType property set to None, Test, or Real. For real signing, the conditional logic checks if we are running a microbuild build as well as checking for SignType == Real. This change does not affect which certs are being used (10006 and 72 for vsix inputs and 160 for the vsix itself), but Trevor suggested changing some of the certs, which can be done at a later time. (changeset 1285798)