PERF: SyntaxFacts.EqualityComparer
Introduce SyntaxFacts.EqualityComparer for both VB and C#.
The framework's EqualityComparer<T>.Default is specialized for enums, but only if the underlying type is an int. Otherwise you get ObjectEqualityComparer<T> which, for enums is pretty bad for perf (lots of boxing). This change introduces a custom comparer, exposed via SyntaxFacts that avoids that problem. (changeset 1285869)