Rationalize our MatchesKind* and IsKind helpers in the VB side of stuff.
History: A long long time ago, the MatchesKind* helpers were added for checking the kinds of tokens. There were even two variants: one that would allow it to be called on nothing, and one that didn't. Then, we merged our syntax models and so the public IsKind methods were added. Also, the ones that should have thrown didn't, so we had three extension methods that all did the same thing.

This is a silly situation so this change does the following:

1) we use the public IsKind helpers whenever possible
2) we rename the private helpers that allow for multiple types to IsKind so it's consistent
3) we delete SyntaxToken.IsParentKind since you can just write SyntaxToken.Parent.IsKind.
4) we delete other helpers that have no users (changeset 1295518)